Add relation to creator
* This is used in spells with effect summon where the summoned creature
doesn't get a *summonedby* flag and so relation to owner is not working.
This way the unit get relation to the master who created it.
* World case where I'm going to use this is summoning Oculus drakes
* Blizzard seems to handle such cases with special spells in dbc that
are missing and sniff gives nothing for them.  The number of spells is
unknown, but existence of atleast 10 is seen. They are used usually as
bonus effects to make the check however mechanic is unknown. Guessing it
is good as adding this.